Move convert function to base utils.py
